pyproject.toml
[tool.poetry]
name = "pe_test"
version = "0.1.0"
description = "Users manager"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
Flask-SQLAlchemy = "^2.5.1"
python-dotenv = "^0.19.0"
psycopg2 = "^2.9.1"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.26.1"
gunicorn = "^20.1.0"
Flask-Mail = "^0.9.1"
PyJWT = "^2.1.0"
Flask-Caching = "^1.10.1"
redis = "^3.5.3"
celery = "^5.1.2"
environs = "^9.3.3"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
